requirements.txt
Django==2.0.1
django-countries==5.1.1
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
